pnsBase-fastApi/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Redis for job management
redis==5.0.1

# HTTP client for API calls
httpx==0.25.2

# Google BigQuery (with dummy config - REPLACE WITH ACTUAL CREDENTIALS)
google-cloud-bigquery==3.13.0

# Existing dependencies from Streamlit project
streamlit==1.28.1
pandas==2.1.3
openai==1.3.5
langchain==0.0.335
langgraph==0.0.20
python-dotenv==1.0.0

# Additional utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
typing-extensions==4.8.0

# Logging
loguru==0.7.2
